bugfix: Fixing contact section
