--- .vercel/output/_functions/chunks/Layout_DJ_-zwXy.mjs
@@ -1,48 +1,0 @@
-import { f as createComponent, m as maybeRenderHead, l as renderScript, r as renderTemplate, e as createAstro, k as renderComponent, n as renderSlot, o as renderHead, h as addAttribute } from './astro/server_fVTS04Ou.mjs';
-import 'kleur/colors';
-import 'clsx';
-/* empty css                           */
-
-const $$Header = createComponent(($$result, $$props, $$slots) => {
-  return renderTemplate`${maybeRenderHead()}<header class="w-full bg-transparent fixed top-0 left-0 right-0 z-[9999]"> <div class="container mx-auto px-6 py-4"> <div class="bg-light-gray/20 rounded-full px-6 py-2 flex justify-between items-center"> <!-- Logo --> <a href="/" class="flex items-center gap-2" aria-label="Ir a la página de inicio de Colinas de la Paz"> <img src="/images/logo.webp" alt="Colinas de la Paz" class="w-auto h-10"> </a> <!-- Desktop Navigation --> <nav class="hidden lg:flex items-center bg-white rounded-full px-2 py-1"> <a href="/" class="text-sm text-dark-gray hover:text-green-primary hover:scale-110 transition-all duration-300 px-4 py-2 rounded-full" aria-label="Ir a la página de inicio">Inicio</a> <div class="relative group"> <button class="text-sm text-dark-gray hover:text-green-primary hover:scale-110 transition-all duration-300 flex items-center gap-1 px-4 py-2 rounded-full" aria-label="Menú de planes" aria-expanded="false">
-Planes
-<svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div class="absolute top-full left-0 mt-2 w-48 bg-white rounded-lg shadow-lg border opacity-0 invisible group-hover:opacity-100 group-hover:visible transition-all duration-200"> <a href="/preneed" class="block px-4 py-3 text-sm text-dark-gray hover:text-green-primary hover:scale-105 transition-all duration-300 first:rounded-t-lg" aria-label="Ver planes de prenecesidad">Prenecesidad</a> <a href="/need" class="block px-4 py-3 text-sm text-dark-gray hover:text-green-primary hover:scale-105 transition-all duration-300 last:rounded-b-lg" aria-label="Ver planes de necesidad inmediata">Necesidad Inmediata</a> </div> </div> <div class="relative group"> <button class="text-sm text-dark-gray hover:text-green-primary hover:scale-110 transition-all duration-300 flex items-center gap-1 px-4 py-2 rounded-full" aria-label="Menú de servicios" aria-expanded="false">
-Servicios
-<svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path> </svg> </button> <div class="absolute top-full left-0 mt-2 w-48 bg-white rounded-lg shadow-lg border opacity-0 invisible group-hover:opacity-100 group-hover:visible transition-all duration-200"> <a href="/funservices" class="block px-4 py-3 text-sm text-dark-gray hover:text-green-primary hover:scale-105 transition-all duration-300 first:rounded-t-lg" aria-label="Ver servicios funerarios">Servicios Funerarios</a> <a href="/slots" class="block px-4 py-3 text-sm text-dark-gray hover:text-green-primary hover:scale-105 transition-all duration-300" aria-label="Ver alquiler de lotes">Alquiler de Lotes</a> <a href="/burning" class="block px-4 py-3 text-sm text-dark-gray hover:text-green-primary hover:scale-105 transition-all duration-300" aria-label="Ver servicios de cremación">Cremación</a> <a href="/optional" class="block px-4 py-3 text-sm text-dark-gray hover:text-green-primary hover:scale-105 transition-all duration-300" aria-label="Ver servicios opcionales">Servicios Opcionales</a> </div> </div> <!-- <a href="/slots" class="text-sm text-dark-gray hover:text-green-primary hover:scale-110 transition-all duration-300 px-4 py-2 rounded-full">Obituarios</a> --> <a href="/ponline" class="text-sm text-dark-gray hover:text-green-primary hover:scale-110 transition-all duration-300 px-4 py-2 rounded-full" aria-label="Acceder a pagos en línea">Pagos en línea</a> <a href="/faq" class="text-sm text-dark-gray hover:text-green-primary hover:scale-110 transition-all duration-300 px-4 py-2 rounded-full" aria-label="Ver preguntas frecuentes">Preguntas frecuentes</a> <a href="/aboutus" class="text-sm text-dark-gray hover:text-green-primary hover:scale-110 transition-all duration-300 px-4 py-2 rounded-full" aria-label="Conocer sobre nosotros">Nosotros</a> </nav> <!-- Contact Button --> <a href="/contact" class="hidden lg:block text-sm font-medium text-green-primary border border-green-primary rounded-full px-6 py-2 hover:bg-green-primary hover:text-white transition-all duration-300" aria-label="Ir a la página de contacto">
-Contacto
-</a> <!-- Mobile Menu Button --> <button id="mobile-menu-btn" class="lg:hidden text-dark-gray p-2" aria-label="Abrir menú de navegación móvil" aria-expanded="false"> <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 6h16M4 12h16m-7 6h7"></path> </svg> </button> </div> <!-- Mobile Menu --> <div id="mobile-menu" class="lg:hidden hidden bg-white/95 backdrop-blur-sm rounded-lg mt-4 p-4 shadow-lg border"> <nav class="space-y-3"> <a href="/" class="block text-sm text-dark-gray hover:text-green-primary transition-colors py-2 px-3 rounded-md">Inicio</a> <!-- Planes Mobile --> <div class="space-y-2"> <div class="text-sm font-medium text-dark-gray py-2 px-3">Planes</div> <a href="/preneed" class="block text-sm text-dark-gray/70 hover:text-green-primary transition-colors py-2 px-6 rounded-md">Prenecesidad</a> <a href="/need" class="block text-sm text-dark-gray/70 hover:text-green-primary transition-colors py-2 px-6 rounded-md">Necesidad Inmediata</a> </div> <!-- Servicios Mobile --> <div class="space-y-2"> <div class="text-sm font-medium text-dark-gray py-2 px-3">Servicios</div> <a href="/funservices" class="block text-sm text-dark-gray/70 hover:text-green-primary transition-colors py-2 px-6 rounded-md">Servicios Funerarios</a> <a href="/slots" class="block text-sm text-dark-gray/70 hover:text-green-primary transition-colors py-2 px-6 rounded-md">Alquiler de Lotes</a> <a href="/burning" class="block text-sm text-dark-gray/70 hover:text-green-primary transition-colors py-2 px-6 rounded-md">Cremación</a> <a href="/optional" class="block text-sm text-dark-gray/70 hover:text-green-primary transition-colors py-2 px-6 rounded-md">Servicios Opcionales</a> </div> <a href="/ponline" class="block text-sm text-dark-gray hover:text-green-primary transition-colors py-2 px-3 rounded-md">Pagos en línea</a> <a href="/faq" class="block text-sm text-dark-gray hover:text-green-primary transition-colors py-2 px-3 rounded-md">Preguntas frecuentes</a> <a href="/aboutus" class="block text-sm text-dark-gray hover:text-green-primary transition-colors py-2 px-3 rounded-md">Nosotros</a> <!-- Contact Button Mobile --> <a href="/contact" class="block text-center text-sm font-medium text-white bg-green-primary rounded-full px-6 py-3 hover:bg-green-dark transition-all duration-300 mt-4">
-Contacto
-</a> </nav> </div> </div> ${renderScript($$result, "C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/Shared/header/Header.astro?astro&type=script&index=0&lang.ts")} </header>  `;
-}, "C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/Shared/header/Header.astro", void 0);
-
-const $$Footer = createComponent(($$result, $$props, $$slots) => {
-  return renderTemplate`${maybeRenderHead()}<footer class="bg-white pt-16 pb-8"> <div class="container mx-auto px-6"> <!-- Footer Main Content --> <div class="grid grid-cols-1 lg:grid-cols-2 gap-8 mb-12"> <!-- Left Column: Logo and Description --> <div class="bg-light-gray/20 rounded-lg p-6 md:p-12 text-center flex flex-col justify-center items-center min-h-full"> <a href="/" class="inline-block mb-6 md:mb-8"> <img src="/images/logo.webp" alt="Colinas de la Paz" class="w-auto h-16 mx-auto"> </a> <p class="text-sm md:text-base text-gray-accessible leading-relaxed mb-6 md:mb-10">
-Cementerio y funeraria privada ubicado en Hato Montaña, Arraiján, Corregimiento Juan Demóstenes, Panamá Oeste, República de Panamá.
-</p> <a href="https://maps.app.goo.gl/wsZpudX7s6Gnitd19" target="_blank" rel="noopener noreferrer" class="text-sm md:text-base font-medium text-dark-gray border border-dark-gray rounded-md px-6 md:px-8 py-2 md:py-3 hover:bg-green-dark hover:text-white transition-all duration-300 inline-flex items-center gap-2 group" aria-label="Ver ubicación de Colinas de la Paz en Google Maps">
-Conoce nuestra ubicación
-<svg class="w-4 md:w-5 h-4 md:h-5 transition-transform duration-300 group-hover:-rotate-90" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M7 7L17 17M17 17H7M17 17V7"></path> </svg> </a> </div> <!-- Right Column: Navigation Sections --> <div class="bg-light-gray/20 rounded-lg p-6 md:p-10"> <div class="space-y-8"> <!-- Primera Fila: Accesos en 3 columnas --> <div> <h4 class="font-bold text-dark-gray mb-6 text-base text-center">Accesos</h4> <div class="grid grid-cols-2 md:grid-cols-3 gap-4 md:gap-6"> <!-- Columna 1 --> <div> <ul class="space-y-2"> <li><a href="/" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Inicio</a></li> <li><a href="/preneed" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Prenecesidad</a></li> <li><a href="/need" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Necesidad Inmediata</a></li> <li><a href="/funservices" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Servicios Funerarios</a></li> </ul> </div> <!-- Columna 2 --> <div> <ul class="space-y-2"> <li><a href="/slots" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Alquiler de Lotes</a></li> <li><a href="/burning" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Cremación</a></li> <li><a href="/optional" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Servicios Opcionales</a></li> <!-- <li><a href="/slots" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Obituarios</a></li> --> <li><a href="/ponline" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Pagos en línea</a></li> </ul> </div> <!-- Columna 3 --> <div> <ul class="space-y-2"> <li><a href="/faq" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Preguntas frecuentes</a></li> <li><a href="/aboutus" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Nosotros</a></li> <li><a href="/contact" class="text-sm text-gray-accessible hover:text-green-dark transition-colors">Contacto</a></li> </ul> </div> </div> </div> <!-- Segunda Fila: Contacto --> <div> <h4 class="font-bold text-dark-gray mb-4 text-base">Contacto</h4> <div class="grid grid-cols-1 md:grid-cols-2 gap-4 md:gap-6"> <div> <h5 class="text-sm font-medium text-dark-gray mb-2">Ventas</h5> <div class="flex items-center gap-2 mb-1"> <div class="w-4 h-4 bg-green-dark rounded-full flex items-center justify-center"> <svg class="w-2.5 h-2.5 text-white" fill="currentColor" viewBox="0 0 24 24"> <path d="M17.472 14.382c-.297-.149-1.758-.867-2.03-.967-.273-.099-.471-.148-.67.15-.197.297-.767.966-.94 1.164-.173.199-.347.223-.644.075-.297-.15-1.255-.463-2.39-1.475-.883-.788-1.48-1.761-1.653-2.059-.173-.297-.018-.458.13-.606.134-.133.298-.347.446-.52.149-.174.198-.298.298-.497.099-.198.05-.371-.025-.52-.075-.149-.669-1.612-.916-2.207-.242-.579-.487-.5-.669-.51-.173-.008-.371-.01-.57-.01-.198 0-.52.074-.792.372-.272.297-1.04 1.016-1.04 2.479 0 1.462 1.065 2.875 1.213 3.074.149.198 2.096 3.2 5.077 4.487.709.306 1.262.489 1.694.625.712.227 1.36.195 1.871.118.571-.085 1.758-.719 2.006-1.413.248-.694.248-1.289.173-1.413-.074-.124-.272-.198-.57-.347m-5.421 7.403h-.004a9.87 9.87 0 01-5.031-1.378l-.361-.214-3.741.982.998-3.648-.235-.374a9.86 9.86 0 01-1.51-5.26c.001-5.45 4.436-9.884 9.888-9.884 2.64 0 5.122 1.03 6.988 2.898a9.825 9.825 0 012.893 6.994c-.003 5.45-4.437 9.884-9.885 9.884m8.413-18.297A11.815 11.815 0 0012.05 0C5.495 0 .16 5.335.157 11.892c0 2.096.547 4.142 1.588 5.945L.057 24l6.305-1.654a11.882 11.882 0 005.683 1.448h.005c6.554 0 11.890-5.335 11.893-11.893A11.821 11.821 0 0020.885 3.787"></path> </svg> </div> <span class="text-sm text-gray-accessible">(+507) 6675-3132</span> </div> <div class="flex items-center gap-2"> <div class="w-4 h-4 bg-green-dark rounded-full flex items-center justify-center"> <svg class="w-2.5 h-2.5 text-white" fill="currentColor" viewBox="0 0 24 24"> <path d="M20 4H4c-1.1 0-1.99.9-1.99 2L2 18c0 1.1.89 2 2 2h16c1.1 0 2-.9 2-2V6c0-1.1-.9-2-2-2zm0 4l-8 5-8-5V6l8 5 8-5v2z"></path> </svg> </div> <span class="text-sm text-gray-accessible">ventas@colinasdelapaz.com</span> </div> </div> <div> <h5 class="text-sm font-medium text-dark-gray mb-2">Traslados (24 horas)</h5> <div class="space-y-1"> <div class="flex items-center gap-2"> <div class="w-4 h-4 bg-green-dark rounded-full flex items-center justify-center"> <svg class="w-2.5 h-2.5 text-white" fill="currentColor" viewBox="0 0 24 24"> <path d="M6.62 10.79c1.44 2.83 3.76 5.14 6.59 6.59l2.2-2.2c.27-.27.67-.36 1.02-.24 1.12.37 2.33.57 3.57.57.55 0 1 .45 1 1V20c0 .55-.45 1-1 1-9.39 0-17-7.61-17-17 0-.55.45-1 1-1h3.5c.55 0 1 .45 1 1 0 1.25.2 2.45.57 3.57.11.35.03.74-.25 1.02l-2.2 2.2z"></path> </svg> </div> <span class="text-sm text-gray-accessible md:hidden">(+507) 6617-1800 / (+507) 6616-3114</span> <span class="text-sm text-gray-accessible hidden md:block">(+507) 6617-1800</span> </div> <div class="hidden md:flex items-center gap-2"> <div class="w-4 h-4 bg-green-dark rounded-full flex items-center justify-center"> <svg class="w-2.5 h-2.5 text-white" fill="currentColor" viewBox="0 0 24 24"> <path d="M6.62 10.79c1.44 2.83 3.76 5.14 6.59 6.59l2.2-2.2c.27-.27.67-.36 1.02-.24 1.12.37 2.33.57 3.57.57.55 0 1 .45 1 1V20c0 .55-.45 1-1 1-9.39 0-17-7.61-17-17 0-.55.45-1 1-1h3.5c.55 0 1 .45 1 1 0 1.25.2 2.45.57 3.57.11.35.03.74-.25 1.02l-2.2 2.2z"></path> </svg> </div> <span class="text-sm text-gray-accessible">(+507) 6616-3114</span> </div> </div> </div> </div> </div> <!-- Tercera Fila: Recursos y Redes Sociales --> <div class="grid grid-cols-1 md:grid-cols-2 gap-4 md:gap-8"> <!-- Recursos --> <div> <h4 class="font-bold text-dark-gray mb-4 text-base">Recursos</h4> <ul class="space-y-2"> <li><a href="#" class="text-sm text-gray-accessible hover:text-green-dark transition-colors" aria-label="Ver política de privacidad de Colinas de la Paz">Política & Privacidad</a></li> <li><a href="#" class="text-sm text-gray-accessible hover:text-green-dark transition-colors" aria-label="Ver términos y condiciones de servicio">Términos & Condiciones</a></li> </ul> </div> <!-- Redes Sociales --> <div> <h4 class="font-bold text-green-dark mb-4 text-base">Síguenos en nuestras redes sociales</h4> <div class="flex gap-3"> <a href="http://instagram.com/colinasdelapazpanama" target="_blank" rel="noopener noreferrer" class="w-8 h-8 bg-white rounded-md flex items-center justify-center hover:bg-green-primary transition-colors border border-green-dark" aria-label="Visitar perfil de Instagram de Colinas de la Paz"> <svg class="w-4 h-4 text-green-dark" fill="currentColor" viewBox="0 0 24 24"><path d="M7.8,2H16.2C19.4,2 22,4.6 22,7.8V16.2A5.8,5.8 0 0,1 16.2,22H7.8C4.6,22 2,19.4 2,16.2V7.8A5.8,5.8 0 0,1 7.8,2M7.6,4A3.6,3.6 0 0,0 4,7.6V16.4C4,18.39 5.61,20 7.6,20H16.4A3.6,3.6 0 0,0 20,16.4V7.6C20,5.61 18.39,4 16.4,4H7.6M17.25,5.5A1.25,1.25 0 0,1 18.5,6.75A1.25,1.25 0 0,1 17.25,8A1.25,1.25 0 0,1 16,6.75A1.25,1.25 0 0,1 17.25,5.5M12,7A5,5 0 0,1 17,12A5,5 0 0,1 12,17A5,5 0 0,1 7,12A5,5 0 0,1 12,7M12,9A3,3 0 0,0 9,12A3,3 0 0,0 12,15A3,3 0 0,0 15,12A3,3 0 0,0 12,9Z"></path></svg> </a> <a href="https://www.facebook.com/colinasdelapazpanama" target="_blank" rel="noopener noreferrer" class="w-8 h-8 bg-white rounded-md flex items-center justify-center hover:bg-green-primary transition-colors border border-green-dark" aria-label="Visitar página de Facebook de Colinas de la Paz"> <svg class="w-4 h-4 text-green-dark" fill="currentColor" viewBox="0 0 24 24"><path d="M12 2.04C6.5 2.04 2 6.53 2 12.06C2 17.06 5.66 21.21 10.44 21.96V14.96H7.9V12.06H10.44V9.85C10.44 7.32 11.93 5.96 14.22 5.96C15.31 5.96 16.45 6.15 16.45 6.15V8.62H15.19C13.95 8.62 13.56 9.39 13.56 10.18V12.06H16.34L15.89 14.96H13.56V21.96A10,10 0 0,0 12,2.04Z"></path></svg> </a> </div> </div> </div> </div> </div> </div> <!-- Copyright Section --> <div class="border-t border-dark-gray/20 pt-6 text-center"> <p class="text-sm text-dark-gray/60">
-Todos los derechos reservados Colinas de la Paz, S.A.
-</p> </div> </div> </footer>`;
-}, "C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/Shared/footer/Footer.astro", void 0);
-
-var __freeze = Object.freeze;
-var __defProp = Object.defineProperty;
-var __template = (cooked, raw) => __freeze(__defProp(cooked, "raw", { value: __freeze(cooked.slice()) }));
-var _a;
-const $$Astro = createAstro("https://colinasdelapaz.com");
-const $$Layout = createComponent(($$result, $$props, $$slots) => {
-  const Astro2 = $$result.createAstro($$Astro, $$props, $$slots);
-  Astro2.self = $$Layout;
-  const {
-    title,
-    description = "Colinas de la Paz - Cementerio y servicios funerarios en Panam\xE1 Oeste. Lotes, nichos, cremaciones y servicios completos con m\xE1s de 15 a\xF1os de experiencia.",
-    image = "/images/og-image.webp",
-    canonical,
-    noindex = false
-  } = Astro2.props;
-  const canonicalURL = canonical || new URL(Astro2.url.pathname, Astro2.site);
-  const imageURL = new URL(image, Astro2.site);
-  return renderTemplate(_a || (_a = __template(['<html lang="es" data-astro-cid-sckkx6r4> <head><meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0"><link rel="icon" type="image/svg+xml" href="/favicon.svg"><meta name="generator"', "><!-- SEO Meta Tags --><title>", '</title><meta name="description"', '><link rel="canonical"', ">", '<!-- Open Graph --><meta property="og:type" content="website"><meta property="og:title"', '><meta property="og:description"', '><meta property="og:image"', '><meta property="og:url"', '><meta property="og:site_name" content="Colinas de la Paz"><meta property="og:locale" content="es_PA"><!-- Twitter Card --><meta name="twitter:card" content="summary_large_image"><meta name="twitter:title"', '><meta name="twitter:description"', '><meta name="twitter:image"', '><!-- Additional SEO --><meta name="theme-color" content="#81BA27"><meta name="author" content="Colinas de la Paz"><meta name="language" content="Spanish"><meta name="geo.region" content="PA"><meta name="geo.placename" content="Panam\xE1 Oeste"><!-- PWA Manifest --><link rel="manifest" href="/manifest.json"><!-- Apple Touch Icons --><link rel="apple-touch-icon" href="/images/logo.webp"><meta name="mobile-web-app-capable" content="yes"><meta name="apple-mobile-web-app-status-bar-style" content="default"><!-- Preload critical resources --><link rel="preload" href="/fonts/Gilroy-Light.woff2" as="font" type="font/woff2" crossorigin="anonymous"><link rel="preload" href="/fonts/Gilroy-Extrabold.woff2" as="font" type="font/woff2" crossorigin="anonymous"><!-- Service Worker Registration -->', '<!-- Structured Data --><script type="application/ld+json">\n		{\n			"@context": "https://schema.org",\n			"@type": "Organization",\n			"name": "Colinas de la Paz",\n			"description": "Cementerio y servicios funerarios en Panam\xE1 Oeste con m\xE1s de 15 a\xF1os de experiencia",\n			"url": "https://colinasdelapaz.com",\n			"logo": "https://colinasdelapaz.com/images/logo.webp",\n			"address": {\n				"@type": "PostalAddress",\n				"addressRegion": "Panam\xE1 Oeste",\n				"addressCountry": "PA"\n			},\n			"contactPoint": {\n				"@type": "ContactPoint",\n				"contactType": "customer service",\n				"availableLanguage": "Spanish"\n			},\n			"sameAs": [\n				"https://www.facebook.com/colinasdelapazpanama",\n				"http://instagram.com/colinasdelapazpanama"\n			],\n			"openingHours": [\n				"Mo-Fr 08:00-17:00",\n				"Sa 08:00-15:00"\n			],\n			"priceRange": "$$",\n			"areaServed": "Panam\xE1 Oeste"\n		}\n		<\/script>', '</head> <body class="bg-white" data-astro-cid-sckkx6r4> ', ' <main class="pt-20" data-astro-cid-sckkx6r4> ', " </main> ", " </body></html>"])), addAttribute(Astro2.generator, "content"), title, addAttribute(description, "content"), addAttribute(canonicalURL, "href"), noindex && renderTemplate`<meta name="robots" content="noindex, nofollow">`, addAttribute(title, "content"), addAttribute(description, "content"), addAttribute(imageURL, "content"), addAttribute(canonicalURL, "content"), addAttribute(title, "content"), addAttribute(description, "content"), addAttribute(imageURL, "content"), renderScript($$result, "C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/layouts/Layout.astro?astro&type=script&index=0&lang.ts"), renderHead(), renderComponent($$result, "Header", $$Header, { "data-astro-cid-sckkx6r4": true }), renderSlot($$result, $$slots["default"]), renderComponent($$result, "Footer", $$Footer, { "data-astro-cid-sckkx6r4": true }));
-}, "C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/layouts/Layout.astro", void 0);
-
-export { $$Layout as $ };
