+++ .vercel/output/_functions/manifest_BsEgltPs.mjs
@@ -1,0 +1,102 @@
+import 'kleur/colors';
+import { p as decodeKey } from './chunks/astro/server_fVTS04Ou.mjs';
+import 'clsx';
+import 'cookie';
+import { N as NOOP_MIDDLEWARE_FN } from './chunks/astro-designed-error-pages_BTXM47jy.mjs';
+import 'es-module-lexer';
+
+function sanitizeParams(params) {
+  return Object.fromEntries(
+    Object.entries(params).map(([key, value]) => {
+      if (typeof value === "string") {
+        return [key, value.normalize().replace(/#/g, "%23").replace(/\?/g, "%3F")];
+      }
+      return [key, value];
+    })
+  );
+}
+function getParameter(part, params) {
+  if (part.spread) {
+    return params[part.content.slice(3)] || "";
+  }
+  if (part.dynamic) {
+    if (!params[part.content]) {
+      throw new TypeError(`Missing parameter: ${part.content}`);
+    }
+    return params[part.content];
+  }
+  return part.content.normalize().replace(/\?/g, "%3F").replace(/#/g, "%23").replace(/%5B/g, "[").replace(/%5D/g, "]");
+}
+function getSegment(segment, params) {
+  const segmentPath = segment.map((part) => getParameter(part, params)).join("");
+  return segmentPath ? "/" + segmentPath : "";
+}
+function getRouteGenerator(segments, addTrailingSlash) {
+  return (params) => {
+    const sanitizedParams = sanitizeParams(params);
+    let trailing = "";
+    if (addTrailingSlash === "always" && segments.length) {
+      trailing = "/";
+    }
+    const path = segments.map((segment) => getSegment(segment, sanitizedParams)).join("") + trailing;
+    return path || "/";
+  };
+}
+
+function deserializeRouteData(rawRouteData) {
+  return {
+    route: rawRouteData.route,
+    type: rawRouteData.type,
+    pattern: new RegExp(rawRouteData.pattern),
+    params: rawRouteData.params,
+    component: rawRouteData.component,
+    generate: getRouteGenerator(rawRouteData.segments, rawRouteData._meta.trailingSlash),
+    pathname: rawRouteData.pathname || void 0,
+    segments: rawRouteData.segments,
+    prerender: rawRouteData.prerender,
+    redirect: rawRouteData.redirect,
+    redirectRoute: rawRouteData.redirectRoute ? deserializeRouteData(rawRouteData.redirectRoute) : void 0,
+    fallbackRoutes: rawRouteData.fallbackRoutes.map((fallback) => {
+      return deserializeRouteData(fallback);
+    }),
+    isIndex: rawRouteData.isIndex,
+    origin: rawRouteData.origin
+  };
+}
+
+function deserializeManifest(serializedManifest) {
+  const routes = [];
+  for (const serializedRoute of serializedManifest.routes) {
+    routes.push({
+      ...serializedRoute,
+      routeData: deserializeRouteData(serializedRoute.routeData)
+    });
+    const route = serializedRoute;
+    route.routeData = deserializeRouteData(serializedRoute.routeData);
+  }
+  const assets = new Set(serializedManifest.assets);
+  const componentMetadata = new Map(serializedManifest.componentMetadata);
+  const inlinedScripts = new Map(serializedManifest.inlinedScripts);
+  const clientDirectives = new Map(serializedManifest.clientDirectives);
+  const serverIslandNameMap = new Map(serializedManifest.serverIslandNameMap);
+  const key = decodeKey(serializedManifest.key);
+  return {
+    // in case user middleware exists, this no-op middleware will be reassigned (see plugin-ssr.ts)
+    middleware() {
+      return { onRequest: NOOP_MIDDLEWARE_FN };
+    },
+    ...serializedManifest,
+    assets,
+    componentMetadata,
+    inlinedScripts,
+    clientDirectives,
+    routes,
+    serverIslandNameMap,
+    key
+  };
+}
+
+const manifest = deserializeManifest({"hrefRoot":"file:///C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina%20Web%20-%20Colinas%20de%20la%20Paz/colinaswebpage/","cacheDir":"file:///C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina%20Web%20-%20Colinas%20de%20la%20Paz/colinaswebpage/node_modules/.astro/","outDir":"file:///C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina%20Web%20-%20Colinas%20de%20la%20Paz/colinaswebpage/dist/","srcDir":"file:///C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina%20Web%20-%20Colinas%20de%20la%20Paz/colinaswebpage/src/","publicDir":"file:///C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina%20Web%20-%20Colinas%20de%20la%20Paz/colinaswebpage/public/","buildClientDir":"file:///C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina%20Web%20-%20Colinas%20de%20la%20Paz/colinaswebpage/dist/client/","buildServerDir":"file:///C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina%20Web%20-%20Colinas%20de%20la%20Paz/colinaswebpage/dist/server/","adapterName":"@astrojs/vercel","routes":[{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"type":"page","component":"_server-islands.astro","params":["name"],"segments":[[{"content":"_server-islands","dynamic":false,"spread":false}],[{"content":"name","dynamic":true,"spread":false}]],"pattern":"^\\/_server-islands\\/([^/]+?)\\/?$","prerender":false,"isIndex":false,"fallbackRoutes":[],"route":"/_server-islands/[name]","origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"contact/index.html","links":[],"scripts":[],"styles":[],"routeData":{"route":"/contact","isIndex":false,"type":"page","pattern":"^\\/contact\\/?$","segments":[[{"content":"contact","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/contact.astro","pathname":"/contact","prerender":true,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"necesidad-inmediata/index.html","links":[],"scripts":[],"styles":[],"routeData":{"route":"/necesidad-inmediata","isIndex":false,"type":"page","pattern":"^\\/necesidad-inmediata\\/?$","segments":[[{"content":"necesidad-inmediata","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/necesidad-inmediata.astro","pathname":"/necesidad-inmediata","prerender":true,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"index.html","links":[],"scripts":[],"styles":[],"routeData":{"route":"/","isIndex":true,"type":"page","pattern":"^\\/$","segments":[],"params":[],"component":"src/pages/index.astro","pathname":"/","prerender":true,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"type":"endpoint","isIndex":false,"route":"/_image","pattern":"^\\/_image\\/?$","segments":[[{"content":"_image","dynamic":false,"spread":false}]],"params":[],"component":"node_modules/astro/dist/assets/endpoint/generic.js","pathname":"/_image","prerender":false,"fallbackRoutes":[],"origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/aboutus","isIndex":false,"type":"page","pattern":"^\\/aboutus\\/?$","segments":[[{"content":"aboutus","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/aboutus.astro","pathname":"/aboutus","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"route":"/api/contact","isIndex":false,"type":"endpoint","pattern":"^\\/api\\/contact\\/?$","segments":[[{"content":"api","dynamic":false,"spread":false}],[{"content":"contact","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/api/contact.js","pathname":"/api/contact","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/burning","isIndex":false,"type":"page","pattern":"^\\/burning\\/?$","segments":[[{"content":"burning","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/burning.astro","pathname":"/burning","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/faq","isIndex":false,"type":"page","pattern":"^\\/faq\\/?$","segments":[[{"content":"faq","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/faq.astro","pathname":"/faq","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/funservices","isIndex":false,"type":"page","pattern":"^\\/funservices\\/?$","segments":[[{"content":"funservices","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/funservices.astro","pathname":"/funservices","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/optional","isIndex":false,"type":"page","pattern":"^\\/optional\\/?$","segments":[[{"content":"optional","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/optional.astro","pathname":"/optional","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/ponline","isIndex":false,"type":"page","pattern":"^\\/ponline\\/?$","segments":[[{"content":"ponline","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/ponline.astro","pathname":"/ponline","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/preneed","isIndex":false,"type":"page","pattern":"^\\/preneed\\/?$","segments":[[{"content":"preneed","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/preneed.astro","pathname":"/preneed","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"route":"/sitemap.xml","isIndex":false,"type":"endpoint","pattern":"^\\/sitemap\\.xml\\/?$","segments":[[{"content":"sitemap.xml","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/sitemap.xml.ts","pathname":"/sitemap.xml","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[],"styles":[{"type":"external","src":"/_astro/aboutus.DK8k4foK.css"}],"routeData":{"route":"/slots","isIndex":false,"type":"page","pattern":"^\\/slots\\/?$","segments":[[{"content":"slots","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/slots.astro","pathname":"/slots","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}}],"site":"https://colinasdelapaz.com","base":"/","trailingSlash":"ignore","compressHTML":true,"componentMetadata":[["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/aboutus.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/burning.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/contact.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/faq.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/funservices.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/index.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/necesidad-inmediata.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/optional.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/ponline.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/preneed.astro",{"propagation":"none","containsHead":true}],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/pages/slots.astro",{"propagation":"none","containsHead":true}]],"renderers":[],"clientDirectives":[["idle","(()=>{var l=(n,t)=>{let i=async()=>{await(await n())()},e=typeof t.value==\"object\"?t.value:void 0,s={timeout:e==null?void 0:e.timeout};\"requestIdleCallback\"in window?window.requestIdleCallback(i,s):setTimeout(i,s.timeout||200)};(self.Astro||(self.Astro={})).idle=l;window.dispatchEvent(new Event(\"astro:idle\"));})();"],["load","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).load=e;window.dispatchEvent(new Event(\"astro:load\"));})();"],["media","(()=>{var n=(a,t)=>{let i=async()=>{await(await a())()};if(t.value){let e=matchMedia(t.value);e.matches?i():e.addEventListener(\"change\",i,{once:!0})}};(self.Astro||(self.Astro={})).media=n;window.dispatchEvent(new Event(\"astro:media\"));})();"],["only","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).only=e;window.dispatchEvent(new Event(\"astro:only\"));})();"],["visible","(()=>{var a=(s,i,o)=>{let r=async()=>{await(await s())()},t=typeof i.value==\"object\"?i.value:void 0,c={rootMargin:t==null?void 0:t.rootMargin},n=new IntersectionObserver(e=>{for(let l of e)if(l.isIntersecting){n.disconnect(),r();break}},c);for(let e of o.children)n.observe(e)};(self.Astro||(self.Astro={})).visible=a;window.dispatchEvent(new Event(\"astro:visible\"));})();"]],"entryModules":{"\u0000noop-middleware":"_noop-middleware.mjs","\u0000noop-actions":"_noop-actions.mjs","\u0000@astro-page:src/pages/aboutus@_@astro":"pages/aboutus.astro.mjs","\u0000@astro-page:src/pages/api/contact@_@js":"pages/api/contact.astro.mjs","\u0000@astro-page:src/pages/burning@_@astro":"pages/burning.astro.mjs","\u0000@astro-page:src/pages/contact@_@astro":"pages/contact.astro.mjs","\u0000@astro-page:src/pages/faq@_@astro":"pages/faq.astro.mjs","\u0000@astro-page:src/pages/funservices@_@astro":"pages/funservices.astro.mjs","\u0000@astro-page:src/pages/necesidad-inmediata@_@astro":"pages/necesidad-inmediata.astro.mjs","\u0000@astro-page:src/pages/optional@_@astro":"pages/optional.astro.mjs","\u0000@astro-page:src/pages/ponline@_@astro":"pages/ponline.astro.mjs","\u0000@astro-page:src/pages/preneed@_@astro":"pages/preneed.astro.mjs","\u0000@astro-page:src/pages/sitemap.xml@_@ts":"pages/sitemap.xml.astro.mjs","\u0000@astro-page:src/pages/slots@_@astro":"pages/slots.astro.mjs","\u0000@astro-page:src/pages/index@_@astro":"pages/index.astro.mjs","\u0000@astrojs-ssr-virtual-entry":"entry.mjs","\u0000@astro-renderers":"renderers.mjs","\u0000@astro-page:node_modules/astro/dist/assets/endpoint/generic@_@js":"pages/_image.astro.mjs","\u0000@astrojs-ssr-adapter":"_@astrojs-ssr-adapter.mjs","\u0000@astrojs-manifest":"manifest_BsEgltPs.mjs","C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/node_modules/astro/dist/assets/services/sharp.js":"chunks/sharp_ntC17HiB.mjs","@astrojs/react/client.js":"_astro/client.DL-_0xdV.js","C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/layouts/Layout.astro?astro&type=script&index=0&lang.ts":"_astro/Layout.astro_astro_type_script_index_0_lang.BQUSSn-Q.js","C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/services/burning/UrnsVariety.astro?astro&type=script&index=0&lang.ts":"_astro/UrnsVariety.astro_astro_type_script_index_0_lang.0bLCa3Ka.js","C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/contact/Hero.astro?astro&type=script&index=0&lang.ts":"_astro/Hero.astro_astro_type_script_index_0_lang.BBaIcnmi.js","C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/Shared/header/Header.astro?astro&type=script&index=0&lang.ts":"_astro/Header.astro_astro_type_script_index_0_lang.CbC4woJq.js","astro:scripts/before-hydration.js":""},"inlinedScripts":[["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/layouts/Layout.astro?astro&type=script&index=0&lang.ts","\"serviceWorker\"in navigator&&window.addEventListener(\"load\",()=>{navigator.serviceWorker.register(\"/sw.js\").then(e=>{console.log(\"SW registered: \",e)}).catch(e=>{console.log(\"SW registration failed: \",e)})});"],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/services/burning/UrnsVariety.astro?astro&type=script&index=0&lang.ts","let r=0;const a=4,s=document.getElementById(\"urnsCarouselTrack\"),e=document.getElementById(\"urnsPrevBtn\"),t=document.getElementById(\"urnsNextBtn\");function n(){s&&(s.style.transform=`translateX(-${r*256}px)`),i()}function i(){!e||!t||(r===0?(e.disabled=!0,e.classList.remove(\"bg-green-primary\",\"hover:bg-green-dark\",\"text-white\"),e.classList.add(\"bg-light-gray\",\"text-dark-gray\")):(e.disabled=!1,e.classList.remove(\"bg-light-gray\",\"text-dark-gray\"),e.classList.add(\"bg-green-primary\",\"hover:bg-green-dark\",\"text-white\")),r===a-1?(t.disabled=!0,t.classList.remove(\"bg-green-primary\",\"hover:bg-green-dark\",\"text-white\"),t.classList.add(\"bg-light-gray\",\"text-dark-gray\")):(t.disabled=!1,t.classList.remove(\"bg-light-gray\",\"text-dark-gray\"),t.classList.add(\"bg-green-primary\",\"hover:bg-green-dark\",\"text-white\")))}function d(){r<a-1&&(r++,n())}function g(){r>0&&(r--,n())}t&&t.addEventListener(\"click\",d);e&&e.addEventListener(\"click\",g);i();"],["C:/Users/gerenteit/Desktop/Semah/Sistemas/Pagina Web - Colinas de la Paz/colinaswebpage/src/components/Shared/header/Header.astro?astro&type=script&index=0&lang.ts","document.addEventListener(\"DOMContentLoaded\",function(){const e=document.getElementById(\"mobile-menu-btn\"),n=document.getElementById(\"mobile-menu\");e&&n&&(e.addEventListener(\"click\",function(){n.classList.toggle(\"hidden\")}),document.addEventListener(\"click\",function(d){const t=d.target;!e.contains(t)&&!n.contains(t)&&n.classList.add(\"hidden\")}))});"]],"assets":["/_astro/aboutus.DK8k4foK.css","/favicon.svg","/manifest.json","/og-image.webp","/robots.txt","/sw.js","/_headers","/fonts/Gilroy-Extrabold.woff","/fonts/Gilroy-Light.woff","/images/logo.webp","/videos/herovideo1.webm","/_astro/client.DL-_0xdV.js","/_astro/Hero.astro_astro_type_script_index_0_lang.BBaIcnmi.js","/images/about/abouthero.webp","/images/about/whoweare.webp","/images/burning/burn1.webp","/images/burning/burn2.webp","/images/burning/burn3.webp","/images/burning/burnhero.webp","/images/burning/osarios.webp","/images/burning/Swipe1.webp","/images/burning/Swipe2.webp","/images/burning/Swipe3.webp","/images/burning/Swipe4.webp","/images/ponline/pagoenlinea.webp","/images/home/desc1.webp","/images/home/desc2.webp","/images/home/Descripcion photo.webp","/images/home/hero.webp","/images/home/isobg.svg","/images/home/Lotes - card.webp","/images/home/Necesidad Inmediata - card.webp","/images/home/nichos.webp","/images/home/Pre-necesidad - card.webp","/images/home/Servicios Funerarios - card.webp","/images/preneed/asesoria.svg","/images/preneed/closecof.webp","/images/preneed/cuotas.svg","/images/preneed/planing.svg","/images/preneed/precio.svg","/images/preneed/preneedhero.webp","/images/preneed/urna.webp","/images/optional/description.webp","/images/optional/opt1.webp","/images/optional/opt2.webp","/images/optional/opt3.webp","/images/optional/opthero.webp","/services/slots/family.webp","/services/slots/single.webp","/services/slots/slotdesc.webp","/services/funservices/fungrid1.webp","/services/funservices/fungrid2.webp","/services/funservices/fungrid3.webp","/services/funservices/funserviceshero.webp","/services/funservices/otherservices.webp","/images/need/bgasistencia.webp","/images/need/inme2.webp","/images/need/needhero.webp","/contact/index.html","/necesidad-inmediata/index.html","/index.html"],"buildFormat":"directory","checkOrigin":true,"serverIslandNameMap":[],"key":"j/L2IAWSLxl7CubFWQu1z7Xu2LN5IiI6+BRs0gVoVcM="});
+if (manifest.sessionConfig) manifest.sessionConfig.driverModule = null;
+
+export { manifest };
