+++ .vercel/output/_functions/pages/necesidad-inmediata.astro.mjs
@@ -1,0 +1,1 @@
+// Contents removed by Astro as it's used for prerendering only
