--- .vercel/output/functions/_render.func/src/pages/need.astro
@@ -1,14 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
-import Hero from '../components/need/Hero.astro';
-import Description from '../components/need/Description.astro';
-import Services from '../components/need/Services.astro';
-import CTA from '../components/need/CTA.astro';
----
-
-<Layout title="Necesidad Inmediata - Colinas de la Paz">
-  <Hero />
-  <Description />
-  <Services />
-  <CTA />
-</Layout>
