+++ astro.config.mjs
@@ -2,11 +2,15 @@
 import { defineConfig } from 'astro/config';
 import tailwindcss from '@tailwindcss/vite';
 import react from '@astrojs/react';
+import vercel from '@astrojs/vercel';
 
 // https://astro.build/config
 export default defineConfig({
   site: 'https://colinasdelapaz.com',
-  output: 'static',
+  output: 'server',
+  adapter: vercel({
+    webAnalytics: { enabled: false }
+  }),
   vite: {
     plugins: [tailwindcss()]
   },
