+++ src/components/contact/Hero.astro
@@ -243,7 +243,7 @@
       try {
         submitButton.textContent = 'ENVIANDO...';
         submitButton.disabled = true;
-        
+
         const response = await fetch('/api/contact', {
           method: 'POST',
           headers: {
@@ -251,9 +251,9 @@
           },
           body: JSON.stringify(data)
         });
-        
+
         const result = await response.json();
-        
+
         if (result.success) {
           form.reset();
           showSuccessModal();
