+++ src/pages/contact.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/contact/Hero.astro';
 import CTA from '../components/contact/CTA.astro';
