+++ src/pages/index.astro
@@ -1,4 +1,6 @@
 ---
+export const prerender = true;
+
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/Home/hero/Hero.astro';
 import Description from '../components/Home/description/Description.astro';
